requirements.txt
asgiref==3.8.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.15.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.6
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
